requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# PDF Processing
pdfplumber==0.11.4

# LangChain
langchain==0.3.14
langchain-text-splitters==0.3.4

# LLM
google-genai>=1.62.0
openai>=1.59.0

# Embeddings & Vector Store
faiss-cpu>=1.13.0
sentence-transformers==3.3.1

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Logging & Metrics
loguru==0.7.3

# Frontend
streamlit==1.41.1
requests==2.32.3

# Testing
pytest==8.3.4
httpx==0.28.1
pytest-cov==6.0.0
pytest-asyncio==0.25.0
